fix(shortStreakNotificationContent): moving short streak message to inactivity
