+++ Notification/Model/FlossReminders/MessageContent.swift
@@ -1,0 +1,16 @@
+//
+//  MessageContent.swift
+//  Notification
+//
+//  Created by Lucas Migge on 04/03/24.
+//
+
+import Foundation
+
+
+struct MessageContent {
+    var title: String
+    var body: String
+    
+}
+
